service implemets

+++ src/main/java/com/masai/repository/VaccineRegistrationDao.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.masai.model.VaccineRegistration;
 
-public interface VaccineRegistrationDao extends JpaRepository<VaccineRegistration, Integer> {
+public interface VaccineRegistrationDao extends JpaRepository<VaccineRegistration, Long> {
+
+//	s
 
 }
